removed uncessary signin layout
