+++ WiseTrack/app/src/main/java/com/faanggang/wisetrack/StartActivity.java
@@ -11,15 +11,16 @@ import android.view.View;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
-import com.faanggang.wisetrack.userauth.NewUserActivity;
+
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
+import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FirebaseFirestore;
-import com.faanggang.wisetrack.userauth.NewUserActivity;
+
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 import java.util.HashMap;
@@ -44,10 +45,10 @@ public class StartActivity extends AppCompatActivity {
     }
 
     public void menuClick(View view) {
-        if(currentUser == null){
-            Intent intent = new Intent(this, NewUserActivity.class);
-            startActivity(intent);
-        } else {
+        if(currentUser == null) {
+            createNewUser();
+        }
+         else {
             Log.w("USERID", currentUser.getUid());
         }
 
@@ -58,11 +59,25 @@ public class StartActivity extends AppCompatActivity {
 
     public void getIDButton(View view) {
         if(currentUser == null){
-            Intent intent = new Intent(this, NewUserActivity.class);
-            startActivity(intent);
+            createNewUser();
         } else {
             Log.w("USERID", currentUser.getUid());
         }
 
+    }
+
+    public void createNewUser(){
+        mAuth.signInAnonymously()
+                .addOnCompleteListener(this, new OnCompleteListener<AuthResult>() {
+                    @Override
+                    public void onComplete(@NonNull Task<AuthResult> task) {
+                        if (task.isSuccessful()) {
+                            Log.d("SIGNIN", "success");
+                            FirebaseUser user = mAuth.getCurrentUser();
+                        } else {
+                            Log.w("SIGNIN", "failure");
+                        }
+                    }
+                });
     }
 }
