--- WiseTrack/app/src/main/java/com/faanggang/wisetrack/userauth/NewUserActivity.java
@@ -1,50 +1,0 @@
-package com.faanggang.wisetrack.userauth;
-
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.util.Log;
-import android.view.View;
-import android.widget.Toast;
-
-import com.faanggang.wisetrack.MainActivity;
-import com.faanggang.wisetrack.R;
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
-import com.google.firebase.auth.AuthResult;
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
-
-public class NewUserActivity extends AppCompatActivity {
-    // open this if user has no auth stuff
-    private FirebaseAuth mAuth;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_new_user);
-
-        mAuth = FirebaseAuth.getInstance();
-    }
-
-    public void anonymousSignIn(View view) {
-        mAuth.signInAnonymously()
-                .addOnCompleteListener(this, new OnCompleteListener<AuthResult>() {
-            @Override
-            public void onComplete(@NonNull Task<AuthResult> task) {
-                if (task.isSuccessful()) {
-                    Log.d("SIGNIN", "success");
-                    FirebaseUser user = mAuth.getCurrentUser();
-                } else {
-                    Log.w("SIGNIN", "failure");
-                }
-            }
-        });
-        if (mAuth.getCurrentUser() != null) {
-            Intent intent = new Intent(this, MainActivity.class);
-            startActivity(intent);
-        }
-    }
-}
