+++ WiseTrack/app/src/main/res/layout/main_menu.xml
@@ -5,8 +5,66 @@
     android:orientation="vertical">
 
     <Button
+        android:id="@+id/menuProfile_Button"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="View My Profile"
+        android:textColor="#090000"
+        android:textColorHighlight="#050000"
+        android:backgroundTint="@android:color/darker_gray"/>
+
+    <Button
+        android:id="@+id/menuViewExperiments_button"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:backgroundTint="@android:color/darker_gray"
+        android:text="View My Experiments"
+        android:textColor="#090000"
+        android:textColorHighlight="#050000" />
+
+    <Button
         android:id="@+id/menuPublish_button"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="Publish Experiment" />
+        android:text="Publish Experiment"
+        android:textColor="#090000"
+        android:textColorHighlight="#050000"
+        android:backgroundTint="@android:color/darker_gray"/>
+
+    <Button
+        android:id="@+id/menuSearch_button"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="Search for Experiments"
+        android:textColor="#090000"
+        android:textColorHighlight="#050000"
+        android:backgroundTint="@android:color/darker_gray"/>
+
+    <Button
+        android:id="@+id/menuSearchUser_button"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="Search for User Profiles"
+        android:textColor="#090000"
+        android:textColorHighlight="#050000"
+        android:backgroundTint="@android:color/darker_gray"/>
+
+    <Button
+        android:id="@+id/menuScanQR_button"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="Scan QR Code"
+        android:textColor="#090000"
+        android:textColorHighlight="#050000"
+        android:backgroundTint="@android:color/darker_gray"/>
+
+    <Button
+        android:id="@+id/menuScanBar_button"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="Scan Bar Code"
+        android:textColor="#090000"
+        android:textColorHighlight="#050000"
+        android:backgroundTint="@android:color/darker_gray"/>
+
 </LinearLayout>
